+++ src/test/java/uk/ac/standrews/cs/utilities/m_tree/MashStringEditDistanceTest.java
@@ -74,7 +74,7 @@ public class MashStringEditDistanceTest {
         Levenshtein levenshtein = new Levenshtein();
 
         @Override
-        public float distance(String s1, String s2) {
+        public double distance(String s1, String s2) {
 
             return levenshtein.distance(s1, s2);
         }
